fix: move hideDevTools function into product flavors
