+++ app/src/espressoFinal/java/xyz/lbres/trickcalculator/testutils/productHelpers.kt
@@ -1,0 +1,5 @@
+package xyz.lbres.trickcalculator.testutils
+
+fun hideDevToolsButton() {
+    throw Exception("Cannot call hideDevToolsButton from final variant")
+}
